Add files via upload
ArcGIS Pro includes a Presence-Only Prediction (MaxEnt) tool that applies the Maximum Entropy (MaxEnt) correlative machine learning model to predict the presence of a phenomenon. The machine learning model uses given known-presence locations and values of explanatory variables to make a prediction about a particular phenomenon. This dataset teaches, or trains, the model to perform a specific task—in this case, to predict suitable locations. The training dataset for the presence-only prediction model includes presence points and explanatory variables.