Add files via upload
he city of Athens, Greece, has been experiencing extreme heat for many years. The city is working with its Chief Heat Officer to create innovative adaptation plans to battle the effects of this extreme heat. As a GIS analyst for Athens, you have been asked to create a heat risk index (HRI) that prioritizes areas for tree planting as part of an adaptation plan to help cool the city. You will work through all the necessary steps to prepare and combine your data to create an HRI for Athens, Greece.